requirements.txt
google_search_results==2.4.2
protobuf==5.29.1
python-dotenv==1.0.1
serpapi==0.1.5
